Add files via upload
Added my version of the Poker Server using LPokerServer as a starting template and Tic Tac Toe Server from a website as a reference. I know that the methods involving adding cards to the deck, drawing hands for two players, and getting card images for drawn cards work, but I'm not sure if the connection to the clients will work. The only thing the client need to do is grab the images from handOneCards and handTwoCards, go through tradeCards method, then run faceOff Method to get the winner. I'm not sure how to handle the ante phase. I'm going to keep messing with the code and try to do some checks with it. Let me know if anything is wrong with the code or if something isn't working on your end.